Finished implementation

+++ simulation/src/math.rs
@@ -1,42 +1,42 @@
-use std::f64::consts::{FRAC_PI_3, PI};
+use std::f32::consts::{FRAC_PI_3, PI};
 use std::iter::Sum;
 use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 
 // at the time of writing the TAU constant is still unstable
-pub const TAU: f64 = 2.0 * PI;
-pub const FRAC_TAU_3: f64 = 2.0 * FRAC_PI_3;
+pub const TAU: f32 = 2.0 * PI;
+pub const FRAC_TAU_3: f32 = 2.0 * FRAC_PI_3;
 
 /// Get the smaller signed angle from `source` to `target`.
 /// The result is in the range `[-PI, PI)`.
-pub fn smallest_angle_between(source: f64, target: f64) -> f64 {
+pub fn smallest_angle_between(source: f32, target: f32) -> f32 {
     let d = target - source;
     (d + PI).rem_euclid(TAU) - PI
 }
 
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct Vector2D {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 impl Vector2D {
-    pub const fn new(x: f64, y: f64) -> Self {
+    pub const fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 
-    pub fn from_polar(angle: f64, radius: f64) -> Self {
+    pub fn from_polar(angle: f32, radius: f32) -> Self {
         let (sin, cos) = angle.sin_cos();
         Self::new(radius * cos, radius * sin)
     }
 
-    pub fn magnitude_squared(self) -> f64 {
+    pub fn magnitude_squared(self) -> f32 {
         self.x * self.x + self.y * self.y
     }
 
-    pub fn magnitude(self) -> f64 {
+    pub fn magnitude(self) -> f32 {
         self.magnitude_squared().sqrt()
     }
 
-    pub fn clamp_magnitude(self, max: f64) -> Self {
+    pub fn clamp_magnitude(self, max: f32) -> Self {
         let mag = self.magnitude();
         if mag > max {
             self / mag * max
@@ -46,7 +46,7 @@ impl Vector2D {
     }
 
     /// Positive angles measured counter-clockwise from positive x axis.
-    pub fn angle(self) -> f64 {
+    pub fn angle(self) -> f32 {
         self.y.atan2(self.x)
     }
 }
@@ -89,31 +89,31 @@ impl Sub for Vector2D {
     }
 }
 
-impl MulAssign<f64> for Vector2D {
-    fn mul_assign(&mut self, scalar: f64) {
+impl MulAssign<f32> for Vector2D {
+    fn mul_assign(&mut self, scalar: f32) {
         self.x *= scalar;
         self.y *= scalar;
     }
 }
-impl Mul<f64> for Vector2D {
+impl Mul<f32> for Vector2D {
     type Output = Self;
 
-    fn mul(mut self, rhs: f64) -> Self::Output {
+    fn mul(mut self, rhs: f32) -> Self::Output {
         self *= rhs;
         self
     }
 }
 
-impl DivAssign<f64> for Vector2D {
-    fn div_assign(&mut self, scalar: f64) {
+impl DivAssign<f32> for Vector2D {
+    fn div_assign(&mut self, scalar: f32) {
         self.x /= scalar;
         self.y /= scalar;
     }
 }
-impl Div<f64> for Vector2D {
+impl Div<f32> for Vector2D {
     type Output = Self;
 
-    fn div(mut self, rhs: f64) -> Self::Output {
+    fn div(mut self, rhs: f32) -> Self::Output {
         self /= rhs;
         self
     }
@@ -126,14 +126,14 @@ impl Sum for Vector2D {
 }
 
 pub trait WeightedMean<T = Self>: Sized {
-    fn weighted_mean(it: impl Iterator<Item = (T, f64)>) -> Option<Self>;
+    fn weighted_mean(it: impl Iterator<Item = (T, f32)>) -> Option<Self>;
 }
 
 impl<T> WeightedMean for T
 where
-    T: AddAssign + Mul<f64, Output = T> + Div<f64, Output = T> + Copy + Default,
+    T: AddAssign + Mul<f32, Output = T> + Div<f32, Output = T> + Copy + Default,
 {
-    fn weighted_mean(it: impl Iterator<Item = (T, f64)>) -> Option<T> {
+    fn weighted_mean(it: impl Iterator<Item = (T, f32)>) -> Option<T> {
         let (sum, total_weight) = it.fold(
             (T::default(), 0.0),
             |(mut sum, total_weight), (value, weight)| {
@@ -155,7 +155,7 @@ pub trait Mean<T = Self>: Sized {
 
 impl<T> Mean for T
 where
-    T: AddAssign + Sub<Output = T> + Div<f64, Output = T> + Copy + Default,
+    T: AddAssign + Sub<Output = T> + Div<f32, Output = T> + Copy + Default,
 {
     fn mean(it: impl Iterator<Item = T>) -> Option<T> {
         let (avg, count) = it.fold((T::default(), 0.0), |(mut avg, mut count), value| {
